Move futures sync-schema to separate file
