+++ workers/loc.api/sync/schema/sync-schema/index.js
@@ -29,6 +29,7 @@ const symbols = require('./symbols')
 const mapSymbols = require('./map-symbols')
 const inactiveCurrencies = require('./inactive-currencies')
 const inactiveSymbols = require('./inactive-symbols')
+const futures = require('./futures')
 
 const _methodCollMap = new Map([
   [SYNC_API_METHODS.LEDGERS, ledgers],
@@ -52,19 +53,7 @@ const _methodCollMap = new Map([
   [SYNC_API_METHODS.MAP_SYMBOLS, mapSymbols],
   [SYNC_API_METHODS.INACTIVE_CURRENCIES, inactiveCurrencies],
   [SYNC_API_METHODS.INACTIVE_SYMBOLS, inactiveSymbols],
-  [
-    SYNC_API_METHODS.FUTURES,
-    {
-      name: ALLOWED_COLLS.FUTURES,
-      maxLimit: 10000,
-      projection: 'pairs',
-      sort: [['pairs', 1]],
-      hasNewData: false,
-      isSyncRequiredAtLeastOnce: true,
-      type: COLLS_TYPES.PUBLIC_UPDATABLE_ARRAY,
-      model: getModelOf(TABLES_NAMES.FUTURES)
-    }
-  ],
+  [SYNC_API_METHODS.FUTURES, futures],
   [
     SYNC_API_METHODS.CURRENCIES,
     {
